backend/requirements.txt
requests
beautifulsoup4
cohere
qdrant-client
python-dotenv
fastapi
uvicorn
